requirements.txt
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-core==2.4.3
google-cloud-translate==3.20.2
googleapis-common-protos==1.69.2
googletrans==4.0.0rc1
pandas==2.2.3
python-dotenv==1.1.0
requests==2.32.3
SQLAlchemy==2.0.40
translate==3.6.1
